+++ hedera-nft-auction-demo-java-node/src/testIntegration/java/com/hedera/demo/auction/node/test/integration/restapi/GetAuctionIntegrationTest.java
@@ -1,6 +1,5 @@
 package com.hedera.demo.auction.node.test.integration.restapi;
 
-import com.google.errorprone.annotations.Var;
 import com.hedera.demo.auction.node.app.SqlConnectionManager;
 import com.hedera.demo.auction.node.app.api.ApiVerticle;
 import com.hedera.demo.auction.node.app.domain.Auction;
@@ -8,7 +7,7 @@ import com.hedera.demo.auction.node.app.repository.AuctionsRepository;
 import com.hedera.demo.auction.node.test.integration.AbstractIntegrationTest;
 import io.vertx.core.DeploymentOptions;
 import io.vertx.core.Vertx;
-import io.vertx.core.json.JsonArray;
+import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.client.WebClient;
 import io.vertx.ext.web.codec.BodyCodec;
 import io.vertx.junit5.VertxExtension;
@@ -23,13 +22,12 @@ import org.testcontainers.containers.PostgreSQLContainer;
 import java.sql.SQLException;
 import java.util.concurrent.TimeUnit;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @ExtendWith(VertxExtension.class)
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-public class GetAuctionsTest extends AbstractIntegrationTest {
+public class GetAuctionIntegrationTest extends AbstractIntegrationTest {
 
     private PostgreSQLContainer postgres;
     private AuctionsRepository auctionsRepository;
@@ -64,25 +62,22 @@ public class GetAuctionsTest extends AbstractIntegrationTest {
     }
 
     @Test
-    public void getAuctionsTest(VertxTestContext testContext) throws SQLException {
-        @Var Auction auction = testAuctionObject(1);
-        Auction newAuction1 = auctionsRepository.createComplete(auction);
-        auction = testAuctionObject(2);
-        Auction newAuction2 = auctionsRepository.createComplete(auction);
-
-        webClient.get(9005, "localhost", "/v1/auctions/")
-                .as(BodyCodec.buffer())
+    public void getAuctionTest(VertxTestContext testContext) throws SQLException {
+        Auction auction = testAuctionObject(1);
+        Auction newAuction = auctionsRepository.createComplete(auction);
+        webClient.get(9005, "localhost", "/v1/auctions/".concat(String.valueOf(newAuction.getId())))
+                .as(BodyCodec.jsonObject())
                 .send(testContext.succeeding(response -> testContext.verify(() -> {
-                    assertNotNull(response);
-                    JsonArray body = new JsonArray(response.body());
+
+                    assertNotNull(response.body());
+                    JsonObject body = JsonObject.mapFrom(response.body());
                     assertNotNull(body);
-                    assertEquals(2, body.size());
-                    verifyAuction(newAuction1, body.getJsonObject(0));
-                    verifyAuction(newAuction2, body.getJsonObject(1));
+
+                    verifyAuction(newAuction, body);
 
                     auctionsRepository.deleteAllAuctions();
+
                     testContext.completeNow();
                 })));
     }
-
 }
