+++ hedera-nft-auction-demo-java-node/src/testIntegration/java/com/hedera/demo/auction/node/test/integration/restapi/GetAuctionsIntegrationTest.java
@@ -1,5 +1,6 @@
 package com.hedera.demo.auction.node.test.integration.restapi;
 
+import com.google.errorprone.annotations.Var;
 import com.hedera.demo.auction.node.app.SqlConnectionManager;
 import com.hedera.demo.auction.node.app.api.ApiVerticle;
 import com.hedera.demo.auction.node.app.domain.Auction;
@@ -7,7 +8,7 @@ import com.hedera.demo.auction.node.app.repository.AuctionsRepository;
 import com.hedera.demo.auction.node.test.integration.AbstractIntegrationTest;
 import io.vertx.core.DeploymentOptions;
 import io.vertx.core.Vertx;
-import io.vertx.core.json.JsonObject;
+import io.vertx.core.json.JsonArray;
 import io.vertx.ext.web.client.WebClient;
 import io.vertx.ext.web.codec.BodyCodec;
 import io.vertx.junit5.VertxExtension;
@@ -22,12 +23,13 @@ import org.testcontainers.containers.PostgreSQLContainer;
 import java.sql.SQLException;
 import java.util.concurrent.TimeUnit;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @ExtendWith(VertxExtension.class)
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-public class GetAuctionTest extends AbstractIntegrationTest {
+public class GetAuctionsIntegrationTest extends AbstractIntegrationTest {
 
     private PostgreSQLContainer postgres;
     private AuctionsRepository auctionsRepository;
@@ -62,22 +64,25 @@ public class GetAuctionTest extends AbstractIntegrationTest {
     }
 
     @Test
-    public void getAuctionTest(VertxTestContext testContext) throws SQLException {
-        Auction auction = testAuctionObject(1);
-        Auction newAuction = auctionsRepository.createComplete(auction);
-        webClient.get(9005, "localhost", "/v1/auctions/".concat(String.valueOf(newAuction.getId())))
-                .as(BodyCodec.jsonObject())
+    public void getAuctionsTest(VertxTestContext testContext) throws SQLException {
+        @Var Auction auction = testAuctionObject(1);
+        Auction newAuction1 = auctionsRepository.createComplete(auction);
+        auction = testAuctionObject(2);
+        Auction newAuction2 = auctionsRepository.createComplete(auction);
+
+        webClient.get(9005, "localhost", "/v1/auctions/")
+                .as(BodyCodec.buffer())
                 .send(testContext.succeeding(response -> testContext.verify(() -> {
-
-                    assertNotNull(response.body());
-                    JsonObject body = JsonObject.mapFrom(response.body());
+                    assertNotNull(response);
+                    JsonArray body = new JsonArray(response.body());
                     assertNotNull(body);
-
-                    verifyAuction(newAuction, body);
+                    assertEquals(2, body.size());
+                    verifyAuction(newAuction1, body.getJsonObject(0));
+                    verifyAuction(newAuction2, body.getJsonObject(1));
 
                     auctionsRepository.deleteAllAuctions();
-
                     testContext.completeNow();
                 })));
     }
+
 }
